Add value_attr to pie chart nodes
You can add a value_attr setting to the pie and donut charts so that
tooltip plugins such as Bootstrap tooltip can display values when
hovering over a pie slice. This works by adding an attribute (e.g.
'title' or 'data-value') as defined in value_attr to each svg path or
circle created.